+++ src/react/use-theme/literals.ts
@@ -1,0 +1,4 @@
+export const THEME_STORAGE_KEY = "theme";
+export const THEME_STORE_STORAGE_KEY = "themeStore";
+export const THEME_CHANGE_EVENT = "qstd:theme-change";
+